+++ webapp/app/js/core.js
@@ -1069,21 +1069,23 @@ app.security = (function () {
             return getCookie(name);
         },
         logout: function () {
-            app.core.Post(app.setting.basepath + 'Security/Logout', undefined, undefined, false)
+            let tenant = localStorage.getItem("Tenant");
+
+            // Limpiar localStorage
+            localStorage.removeItem('Token');
+            localStorage.removeItem('Username');
+            localStorage.removeItem('Color1Tenant');
+            localStorage.removeItem('Color2Tenant');
+            localStorage.removeItem('Roles');
+            localStorage.removeItem('Expires');
+            localStorage.removeItem('LastActivity');
+            localStorage.removeItem('Navegation');
+
+            app.core.Get(app.setting.basepath + 'Security/Logout?tenant=' + tenant)
                 .done(function (data) {
                     if (data.success) {
-                        // Limpiar localStorage
-                        localStorage.removeItem('Token');
-                        localStorage.removeItem('Username');
-                        localStorage.removeItem('Color1Tenant');
-                        localStorage.removeItem('Color2Tenant');
-                        localStorage.removeItem('Roles');
-                        localStorage.removeItem('Expires');
-                        localStorage.removeItem('LastActivity');
-                        localStorage.removeItem('Navegation');
-
                         // Redirigir al login
-                        window.location.replace(app.setting.basepath + 'Security/Login');
+                        window.location.replace(data.url);
                     } else {
                         toastr.error('Error al cerrar sesión', 'Error');
                     }